backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
neo4j==5.14.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pydantic-settings==2.1.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pdfplumber==0.10.3
lxml==4.9.3
aiofiles==23.2.1

